requirements.txt
absl-py==2.1.0
adjustText==1.1.1
aiohttp==3.9.5
aiolimiter==1.1.0
aiosignal==1.3.1
anndata==0.10.7
annotated-types==0.7.0
anyio==4.4.0
array_api_compat==1.7.1
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.2.0
Babel==2.8.0
bcc==0.18.0
bcrypt==4.1.3
beautifulsoup4==4.10.0
bidict==0.23.1
biopython==1.83
blinker==1.8.2
Brlapi==0.8.3
capstone==5.0.1
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chex==0.1.86
chrome-gnome-shell==0.0.0
click==8.1.7
colabdock @ git+https://github.com/JeffSHF/ColabDock.git@d3d40572cbe09fef151bd08cfc74114ef661b730
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
contextlib2==21.6.0
contourpy==1.2.1
cryptography==41.0.7
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distorm3==3.5.2
distro==1.7.0
dm-haiku==0.0.12
dm-tree==0.1.8
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
etils==1.7.0
exceptiongroup==1.2.1
executing==2.0.1
fair-esm==2.0.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
flake8==7.0.0
Flask==3.0.1
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
flax==0.8.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
future==0.18.2
greenlet==3.0.3
gseapy==1.1.3
h11==0.14.0
h5py==3.11.0
hidpidaemon==18.4.6
html5lib==1.1
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
humanfriendly==10.0
idna==2.7
immutabledict==4.2.0
impacket==0.9.24
importlib-metadata==4.6.4
importlib_resources==6.4.0
ipaddr==2.2.0
ipynb-py-convert==0.4.6
ipython==8.24.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
jmp==0.0.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-path==0.3.3
jsonschema-specifications==2023.12.1
kernelstub==3.1.4
keyring==23.5.0
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.10.0
ldap3==2.8.1
ldapdomaindump==0.9.3
legacy-api-wrap==1.4
libvirt-python==8.0.0
llvmlite==0.42.0
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
ml-collections==0.1.1
ml-dtypes==0.4.0
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
natsort==8.4.0
nest-asyncio==1.6.0
netaddr==0.9.0
netifaces==0.11.0
networkx==3.3
numba==0.59.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.0
offat==0.19.3
onnxruntime==1.18.0
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
openpyxl==3.1.2
opt-einsum==3.3.0
optax==0.2.2
orbax-checkpoint==0.5.20
orjson==3.10.5
ospd==21.4.4
ospd-openvas==21.4.3
ovs==2.17.9
packaging==24.0
pandas==2.2.2
paramiko==3.4.0
parso==0.8.4
pathable==0.4.3
patsy==0.5.6
pefile==2023.2.7
pexpect==4.9.0
pillow==10.3.0
pop-transition==1.1.2
prompt-toolkit==3.0.43
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py3Dmol==2.1.0
py3dns==4.0.0
pyasn1==0.4.8
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.22
pycrypto==2.6.1
pycryptodomex==3.11.0
pycups==2.0.1
pydantic==2.7.4
pydantic_core==2.18.4
pydbus==0.6.0
pydeseq2==0.4.9
pyflakes==3.2.0
Pygments==2.18.0
PyGObject==3.42.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.8.0
PyNaCl==1.5.0
pyngrok==7.1.6
pynndescent==0.5.12
pyOpenSSL==23.2.0
pyparsing==2.4.7
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-engineio==4.9.0
python-gnupg==0.4.8
python-multipart==0.0.9
python-owasp-zap-v2.4==0.0.14
python-socketio==5.11.2
python-xlib==0.29
pytz==2024.1
pyxdg==0.27
PyYAML==6.0.1
redis==3.5.3
referencing==0.35.1
repolib==2.2.1
repoman==1.4.0
requests==2.31.0
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.20.0
sanbomics==0.1.0
scanpy==1.10.1
scikit-learn==1.5.0
scipy==1.13.1
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
session-info==1.0.0
sessioninstaller==0.0.0
shellingham==1.5.4
simple-websocket==1.0.0
simplejson==3.17.6
six==1.10.0
sniffio==1.3.1
sortedcontainers==2.1.0
SoundCard==0.4.3
soundfile==0.12.1
soupsieve==2.3.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.37.2
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.12.1
systemd-python==234
tabulate==0.7.5
tenacity==8.5.0
tensorstore==0.1.63
termcolor==1.1.0
terminable-thread==0.7.1
terminaltables==3.1.0
texttable==1.6.7
threadpoolctl==3.5.0
toolz==0.12.1
torch==2.3.1
tqdm==4.66.4
traitlets==5.14.3
triton==2.3.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.9.0
umap-learn==0.5.6
urllib3==1.24.3
uvicorn==0.30.1
uvloop==0.19.0
vboxapi==1.0
volatility3 @ git+https://github.com/volatilityfoundation/volatility3.git@1b3ba6ab2ef6f086993ad382f68d9f01a4ae7bdf
wadllib==1.3.6
watchfiles==0.22.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.3
wrapt==1.13.3
wsproto==1.2.0
xdg==5
xkit==0.0.0
yara==1.7.7
yarl==1.9.4
zapcli==0.10.0
zipp==1.0.0
